Changed organization from a seperate key and offset class to a combined one
